GameMap, GameOver scene
GameMap & GameOver

+++ Assets/Scripts/Necessary/MainScene.cs
@@ -7,6 +7,7 @@ public class MainScene : MonoBehaviour {
 	public UFOEnemy UFOenemy;
 	public GameObject enemybody; 
 	public GameObject playerbody; 
+	public GameObject BurstButton;
 	public GameObject SkillMenu;
 	public GameObject ItemMenu;
 	public GameObject question;
@@ -17,15 +18,18 @@ public class MainScene : MonoBehaviour {
 	private bool isShowing;
 	private bool isShowingSkill;
 	private bool isShowingItem;
+//	private bool isShowingBurst;
 
 	// Use this for initialization
 	void Start () {
 		this.enemyDestroyed = false;
 		this.potionDeployed = false;
 		this.question.SetActive(false);
+		this.BurstButton.SetActive(false);
 		this.SkillMenu.SetActive(false);
 		this.ItemMenu.SetActive(false);
 		this.isShowingSkill = false;
+//		this.isShowingBurst = false;
 		this.isShowing = false;
 	}
 
@@ -94,11 +98,13 @@ public class MainScene : MonoBehaviour {
 		{
 			int attackStart = this.playerOne.attackStart;
 			int attackEnd = this.playerOne.attackEnd;
-			Attack = Random.Range(attackStart, attackEnd);
+			Attack = (Random.Range(attackStart, attackEnd)+10) - (this.UFOenemy.Def);
+			if (Attack <= 0)
+				Attack = 1;
 			this.UFOenemy.health -= Attack;
-			this.UFOenemy.updateSlider();
 			this.playerOne.hasAttacked = true;
 			this.playerOne.playAttackAnim();
+			this.UFOenemy.updateSlider();
 		}
 	}
 	
@@ -144,20 +150,59 @@ public class MainScene : MonoBehaviour {
 		{
 			int attackStart = this.UFOenemy.attackStart;
 			int attackEnd = this.UFOenemy.attackEnd;
-			int Attack = Random.Range(attackStart, attackEnd);
+			int Attack = (Random.Range(attackStart, attackEnd)+10) - (this.playerOne.Def);
+			if (Attack <= 0)
+				Attack = 1;
 			this.playerOne.health -= Attack;
-			this.playerOne.burst += (Attack*50)/this.playerOne.maxHealth;
+			if (Attack >= this.playerOne.maxHealth) //kondisi jika damage > max health
+				this.playerOne.burst += 50;
+			else
+				this.playerOne.burst += (Attack*50)/this.playerOne.maxHealth;
+//			this.playerOne.burst += 100;
 			this.UFOenemy.playAttackAnim();
 			this.playerOne.updateSlider();
 			if (this.playerOne.burst == 100)
 			{
-				this.playerOne.burst = 0;
-				this.playerOne.updateSlider();
-				this.UFOenemy.health -= 200;
-				this.playerOne.playAttackAnim();
-				this.UFOenemy.updateSlider();
+//				this.isShowingBurst = true;
+//				this.BurstButton.SetActive(this.isShowingBurst);
+				this.BurstButton.SetActive(true);
 			}
 			this.playerOne.hasAttacked = false;
+		}
+
+	}
+
+	public void BurstMode()
+	{
+		Debug.Log ("Burst Show");
+		if (this.isShowingItem == true)
+		{
+			this.isShowingItem = false;
+			this.ItemMenu.SetActive(this.isShowingItem);
+		}
+		else if (this.isShowingSkill == true)
+		{
+			this.isShowingSkill = false;
+			this.SkillMenu.SetActive(this.isShowingSkill);
+		}
+		
+		if (this.playerDestroyed == false && this.UFOenemy.attackPlayed() == false && this.enemyDestroyed == false) 
+		{
+			this.BurstButton.SetActive(false);
+			this.playerOne.burst = 0;
+			this.playerOne.updateSlider();
+			this.UFOenemy.health -= 100;
+			this.playerOne.playAttackAnim();
+			this.UFOenemy.updateSlider();
+			//			Debug.Log ("Burst Show");
+//			if (this.isShowingBurst == true)
+//			{
+//				this.isShowingBurst = false;
+//				this.BurstButton.SetActive(this.isShowingBurst);
+//			}
+//			this.isShowingItem = !this.isShowingItem;
+//			this.ItemMenu.SetActive(this.isShowingItem);
+			//			this.checkSkill();
 		}
 
 	}
@@ -184,10 +229,11 @@ public class MainScene : MonoBehaviour {
 	
 	// Update is called once per frame
 	void Update () {
-		if (this.playerDestroyed == true || this.enemyDestroyed == true) 
-		{
-			Application.LoadLevel("GameMenu");
-		}
+//		if (this.playerDestroyed == true || this.enemyDestroyed == true) 
+//		{
+//			AutoFade.LoadLevel("GameMenu", 3,1,Color.black);
+//			Application.LoadLevel("GameOver");
+//		}
 
 		if((this.playerDead() || this.potionDeployed) && this.playerOne.hasAttacked == true)
 		{
@@ -195,16 +241,18 @@ public class MainScene : MonoBehaviour {
 			this.potionDeployed = false;
 		}
 
-		if(this.playerOne.health <= 0)
+		if(this.playerOne.health <= 0 && this.playerDestroyed == false)
 		{
 			this.destroyPlayer();
 			Debug.Log("You Lose!");
+			Application.LoadLevel("GameOver");
 		}
 
 		else if(this.UFOenemy.health <= 0 && this.enemyDestroyed == false)
 		{
 			this.destroyEnemy();
 			Debug.Log("You Win!");
+			Application.LoadLevel("GameMap");
 		}
 
 //		if(this.questionAsked)
